// src/types/modules/v1/health-check/data/health-check-data.type.ts
export interface HealthCheckData {
  cpu: number;
  memory: {
    total: number;
    used: number;
    usedPercent: number;
  };
  disk: Array<{
    fs: string;
    total: number;
    used: number;
    free: number;
    usedPercent: number;
  }>;
  network: Array<{
    iface: string;
    ip4: string;
  }>;
}